requirements.txt
Beaker==1.8.0
Django==1.9.7
Flask==0.11.1
Glances==2.6.1
Jinja2==2.7.2
Mako==1.0.4
MarkupSafe==0.18
MySQL-python==1.2.3
PAM==0.4.2
Pillow==2.3.0
PyYAML==3.11
Pygments==2.1.3
SQLAlchemy==1.0.14
Twisted-Core==13.2.0
Twisted-Web==13.2.0
WebOb==1.6.1
WebTest==2.0.23
Werkzeug==0.11.11
adium-theme-ubuntu==0.3.4
apt-xapian-index==0.45
argparse==1.2.1
backports-abc==0.4
backports.shutil-get-terminal-size==1.0.0
backports.ssl-match-hostname==3.5.0.1
beautifulsoup4==4.5.0
bottle==0.12.9
cached-property==1.3.0
certifi==2016.2.28
chardet==2.0.1
click==6.6
cliff==2.1.0
cmd2==0.6.8
colorama==0.2.5
command-not-found==0.3
configparser==3.3.0.post2
coverage==4.1
debtagshw==0.1
decorator==3.4.0
defer==1.0.6
dirspec==13.10
django-environ==0.4.0
docker-compose==1.8.1
docker-py==1.10.3
docker-pycreds==0.2.1
dockerpty==0.4.1
docopt==0.6.2
duplicity==0.6.23
entrypoints==0.2.1
enum34==1.1.6
environ==1.0
funcsigs==1.0.2
functools32==3.2.3.post2
gprof2dot==2015.12.1
gunicorn==19.6.0
gyp==0.1
html5lib==0.999
httpbin==0.5.0
httplib2==0.8
ipaddress==1.0.17
ipykernel==4.3.1
ipython==1.2.1
ipython-genutils==0.1.0
itsdangerous==0.24
jsonschema==2.5.1
jupyter-client==4.2.2
jupyter-core==4.1.0
lockfile==0.8
logutils==0.3.3
lxml==3.3.3
meld3==0.6.10
mercurial==2.8.2
mistune==0.7.2
nbconvert==4.2.0
nbformat==4.0.1
notebook==4.2.0
oauthlib==0.6.1
oneconf==0.3.7.14.04.1
opensky-api==1.0.0
pathlib2==2.1.0
pbr==1.10.0
pecan==1.1.2
pexpect==3.1
pickleshare==0.7.2
piston-mini-client==0.7.5
prettytable==0.7.2
psutil==4.1.0
ptyprocess==0.5.1
pyOpenSSL==0.13
pycrypto==2.6.1
pycups==1.9.66
pycurl==7.19.3
pygobject==3.12.0
pymongo==3.3.0
pyparsing==2.1.5
pyserial==2.6
pysmbc==1.0.14.1
python-apt==0.9.3.5ubuntu2
python-dateutil==2.5.3
python-debian==0.1.21-nmu2ubuntu2
pyxdg==0.25
pyzmq==14.0.1
records==0.4.3
redis==2.10.5
reportlab==3.0
requests==2.2.1
selenium==2.53.6
sessioninstaller==0.0.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.5.2
smiley==0.8.1
software-center-aptd-plugins==0.0.0
stevedore==1.16.0
supervisor==3.0b2
system-service==0.1.6
tablib==0.11.2
terminado==0.6
texttable==0.8.4
tornado==3.1.1
traitlets==4.2.1
unicodecsv==0.14.1
unity-lens-photos==1.0
urllib3==1.7.1
vboxapi==1.0
virtualenv==1.11.4
virtualenv-clone==0.2.6
virtualenvwrapper==4.7.2
waitress==0.9.0
websocket-client==0.37.0
wheel==0.24.0
wsgiref==0.1.2
xdiagnose==3.6.3build2
zope.interface==4.0.5
